+++ spec/unit/resources/logical_interconnect/update_from_group_spec.rb
@@ -9,4 +9,9 @@ describe 'oneview_test::logical_interconnect_update_from_group' do
     expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:compliance).and_return(true)
     expect(real_chef_run).to update_oneview_logical_interconnect_from_group('LogicalInterconnect-update_from_group')
   end
+
+  it 'fails if the resource is not found' do
+    expect_any_instance_of(OneviewSDK::LogicalInterconnect).to receive(:retrieve!).and_return(false)
+    expect { real_chef_run }.to raise_error(RuntimeError, /not found/)
+  end
 end